Versions 3.2.2 and later are only available at boilerbay.com
The distribution of AirConcurrentMap is now exclusively at
boilerbay.com. Further development and bug fixes are
carried on there, and licensing and support is now a fixed, non-
negotiable price, with purchasing directly on the site.
There are trial and deployment versions there.
No further changes will be checked into github.

We deeply appreciate the many users we have had
up to this point and hope they will understand the
need to become a more commercial enterprise,
so we can provide even more rapid development and
support.

The deployment and trial license agreements for
3.2.2 and later are checked in here.